feat: 재생목록 페이지 playlist 컴포넌트화 (#58)
- playlist를 PlaylistBox로 컴포넌트화
